Factored commonality between NestedPathMonitor and Pathmonitor out into a common base class.  Integrated the use of promises into the handling of this.parse, so that asynchronous tasks, such as pulling body text from s3 (in the case where firebase holds only a reference to the location) into elasticsearc.
