+++ lib/NestedPathMonitor.js
@@ -1,95 +1,86 @@
 
-var fbutil = require('./fbutil');
-var logger = require('./logging').logger;
+'use strict';
+
+var fbutil = require('./fbutil'),
+  logger = require('./logging').logger,
+  AbstractPathMonitor = require('./AbstractPathMonitor');
 
 function NestedPathMonitor(esc, path) {
-  this.ref = fbutil.fbRef(path.path);
-  logger.info('Indexing %s/%s using path "%s"', path.index, path.type, fbutil.pathName(this.ref));
-  this.esc = esc;
-  
-  this.index = path.index;
-  this.type = path.type;
   this.parentType = path.nested.parentType;
   this.parentField = path.nested.parentField;
   this.childIdField = path.nested.childIdField || "id";
-  this.filter = path.filter || function() { return true; };
-  this.parse  = path.parser || function(data) { return parseKeys(data, path.fields, path.omit); };
+
+  AbstractPathMonitor.call(this, esc, path);
+}
+
+NestedPathMonitor.prototype = Object.create(AbstractPathMonitor.prototype);
+NestedPathMonitor.prototype.constructor = NestedPathMonitor;
+
+NestedPathMonitor.prototype._init = function() {
+  var self = this;
+
+  this._preloadExisting().then(function(cursor) {
+
+    self.addListener = self.ref.on('child_added', self._process.bind(self, self._childUpsert));
+    self.changeListener = self.ref.on('child_changed', self._process.bind(self, self._childUpsert));
   
-  this._init();
-}
+    logger.info("Flashlight initialized type " + self.type + "; " + fbutil.getListenerCount() + " listeners.");
+  })
+  .catch(function(err) {
+    logger.error("could not preload existing data in firebase: " + err);
+  });
+};
 
-NestedPathMonitor.prototype = {
-  _init: function() {
-    this.addListener = this.ref.on('child_added', this._process.bind(this, this._childUpsert));
-    this.changeListener = this.ref.on('child_changed', this._process.bind(this, this._childUpsert));
-    
-    logger.info("Flashlight initialized type " + this.type + "; " + fbutil.getListenerCount() + " listeners.");
-  },
+NestedPathMonitor.prototype._stop = function() {
+  this.ref.off('child_added', this.addListener);
+  this.ref.off('child_changed', this.changeListener);
+};
 
-  _stop: function() {
-    this.ref.off('child_added', this.addListener);
-    this.ref.off('child_changed', this.changeListener);
-  },
-
-  _process: function(fn, snap) {
-    var dat = snap.exportVal();
-    if( this.filter(dat) ) {
-      this.parse(dat).then( function(parsed) {
-        fn.call(this, snap.name(), parsed);
-      });
-    }
-  },
-
-  _childUpsert: function(key, data) {
-    var self = this;
-    self.esc.get({
-      index: self.index,
-      type: self.parentType,
-      id: key,
-      _source_include: [ self.parentField ]
-    }).then( function(rsp) {
-      // Because we have to re-index the parent doc, we want to include
-      // any existing nested elements.
-      var refreshed = rsp._source[self.parentField];
-      var allIds = refreshed.map( function(c) { 
-        return c[self.childIdField];
-      });
-      // Overwrite the returned result set with any changed contents
-      // of the snap.  Add any new additions.  There is not presently
-      // a way to identify deletions.
-      data.forEach( function(d) {
-        var idx = allIds.indexOf( d[self.childIdField] );
-        if (idx >= 0) {
-          refreshed[idx] = d;
-        }
-        else {
-          refreshed.push( d );
-        }
-      });
-      // Then write the refreshed list of nested elements into the parent
-      // doc as an "update".
-      var updateObj = {
-        index: self.index, 
-        type: self.parentType, 
-        id: key,
-        body: {
-          doc: { }
-        }
-      };
-      updateObj.body.doc[self.parentField] = refreshed;
-      self.esc.update(updateObj).then( function(rsp) {
-        logger.debug('updated', nameFor(self, key));
-      }).catch( function(err) {
-        logger.error('failed to update %s: %s', nameFor(self, key), err);
-      });
-    }).catch( function(err) {
-      logger.error("Could not index " + self.type + ": " + err);
+NestedPathMonitor.prototype._childUpsert = function(key, data) {
+  var self = this;
+  self.esc.get({
+    index: self.index,
+    type: self.parentType,
+    id: key,
+    _source_include: [ self.parentField ]
+  }).then( function(rsp) {
+    // Because we have to re-index the parent doc, we want to include
+    // any existing nested elements.
+    var refreshed = rsp._source[self.parentField];
+    var allIds = refreshed.map( function(c) { 
+      return c[self.childIdField];
     });
-  }
-};  
-
-function nameFor(path, key) {
-  return path.index + '/' + path.type + '/' + key;
-}
+    // Overwrite the returned result set with any changed contents
+    // of the snap.  Add any new additions.  There is not presently
+    // a way to identify deletions.
+    data.forEach( function(d) {
+      var idx = allIds.indexOf( d[self.childIdField] );
+      if (idx >= 0) {
+        refreshed[idx] = d;
+      }
+      else {
+        refreshed.push( d );
+      }
+    });
+    // Then write the refreshed list of nested elements into the parent
+    // doc as an "update".
+    var updateObj = {
+      index: self.index, 
+      type: self.parentType, 
+      id: key,
+      body: {
+        doc: { }
+      }
+    };
+    updateObj.body.doc[self.parentField] = refreshed;
+    self.esc.update(updateObj).then( function(rsp) {
+      logger.debug('updated', self._nameFor(key));
+    }).catch( function(err) {
+      logger.error('failed to update %s: %s', self._nameFor(key), err);
+    });
+  }).catch( function(err) {
+    logger.error("Could not index " + self.type + ": " + err);
+  });
+};
 
 module.exports = NestedPathMonitor;
