+++ lib/PathMonitor.js
@@ -4,219 +4,83 @@
 var S = require('string'),
    _ = require('lodash'),
    Q = require('q'),
-   Fireproof = require('fireproof'),
    fbutil = require('./fbutil'),
+   AbstractPathMonitor = require('./AbstractPathMonitor'),
    DynamicPathMonitor = require('./DynamicPathMonitor'),
    NestedPathMonitor = require('./NestedPathMonitor'),
    logger = require('./logging').logger;
 
-var PAGE_SIZE = 500;
-
 function PathMonitor(esc, path) {
-  this.ref = fbutil.fbRef(path.path);
-  logger.info('Indexing %s/%s using path "%s"', path.index, path.type, fbutil.pathName(this.ref));
-  this.esc = esc;
-  
-  this.index = path.index;
-  this.type  = path.type;
-  this.filter = path.filter || function() { return true; };
-  this.parse  = path.parser || function(data) { return parseKeys(data, path.fields, path.omit); };
-  
-  this._init();
+  AbstractPathMonitor.call(this, esc, path);
 }
 
-PathMonitor.prototype = {
-  _preloadExisting: function() {
-    var pager;
-    var self = this;
+PathMonitor.prototype = Object.create(AbstractPathMonitor.prototype);
+PathMonitor.prototype.constructor = PathMonitor;
 
-    var lastKey = null;
-    var lastPriority = null;
-
-    return (function nextPage() {
-
-      var promise;
-      if (!pager) {
-        pager = new Fireproof.Pager(self.ref, PAGE_SIZE);
-        promise = pager;
-      } else {
-        promise = pager.next(PAGE_SIZE);
-      }
-      
-      return promise.then(function(snaps) {
-        logger.info('got ' + snaps.length + ' snaps.');
-        if (snaps.length > 0) {
-
-          // This is much gnarlier than I wish it were; we first resolve
-          // the promises on parse() (requried for a use case involving
-          // dereferencing blobs stored in s3), and then bundle all the 
-          // parsed snaps into the elasticsearch.bulk() call.
-          return Q.all(
-            snaps.filter(function(s) { return self.filter(s.val()); })
-              .map(function(snap) {
-                lastKey = snap.name();
-                lastPriority = snap.getPriority();
-                return self.parse( snap.val() ).then( function(value) {
-                  return { id: snap.name(), value: value };
-                })
-              })
-          ).then( function(parsed) {
-          
-            return self.esc.bulk({
-              body: parsed.map( function(p) {
-                return [
-                  {
-                    index: {
-                      _index: self.index, 
-                      _type: self.type, 
-                      _id: p.id, 
-                      _source: true
-                    }
-                  },
-                  p.value
-                ]
-              }).reduce(function(a,b) { return a.concat(b); })
-            }).then(function() {
-              return nextPage();
-            });
-
-          });
-
-        }
-        else {
-          return { key: lastKey, priority: lastPriority };
-        }
-
-      });
-
-    })();
-
-  },
-
-  _init: function() {
-    var self = this;
-
-    this._preloadExisting().then(function(cursor) {
-
-      self.addListener = self.ref.startAt(cursor.priority, cursor.key)
-        .on('child_added', self._process.bind(self, self._childAdded));
-      self.changeListener = self.ref.on('child_changed', self._process.bind(self, self._childChanged));
-      self.removeListener = self.ref.on('child_removed', self._process.bind(self, self._childRemoved));
-      
-      logger.info("Flashlight initialized type " + self.type + "; " + fbutil.getListenerCount() + " listeners.");
-
-    })
-    .catch(function(err) {
-      logger.error("Could not preload existing data in firebase: " + err);
-    });
-  },
-
-  _stop: function() {
-    this.ref.off('child_added', this.addListener);
-    this.ref.off('child_changed', this.changeListener);
-    this.ref.off('child_removed', this.removeListener);
-  },
+PathMonitor.prototype._init = function() {
+  var self = this;
   
-  _process: function(fn, snap) {
-    var dat = snap.val();
-    if( this.filter(dat) ) {
-      this.parse(dat).then( function(parsed) {
-        fn.call(this, snap.name(), parsed);
-      });
-    }
-  },
-  
-  _childAdded: function(key, data) {
-    var name = nameFor(this, key);
-    this.esc.index({
-      index: this.index, 
-      type: this.type, 
-      id: key,
-      body: data
-    }).then( function(rsp) {
-      logger.debug('indexed', name);
-    }).catch( function(err) {
-      logger.error('failed to index %s: %s', name, err);
-    });
-  },
-  
-  _childChanged: function(key, data) {
-    var name = nameFor(this, key);
-    this.esc.index({
-      index: this.index, 
-      type: this.type, 
-      id: key,
-      body: data
-    }).then( function(rsp) {
-      logger.debug('updated', name);
-    }).catch( function(err) {
-      logger.error('failed to update %s: %s', name, err);
-    });
-  },
-  
-  _childRemoved: function(key, data) {
-    var name = nameFor(this, key);
-    this.esc.delete({
-      index: this.index, 
-      type: this.type, 
-      id: key, 
-    }).then( function(rsp) {
-      logger.debug('deleted', name);
-    }).catch( function(err) {
-      logger.error('failed to delete %s: %s', name, err);
-    });
-  }
+  this._preloadExisting().then(function(cursor) {
+    
+    self.addListener = self.ref.startAt(cursor.priority, cursor.key)
+      .on('child_added', self._process.bind(self, self._childAdded));
+    self.changeListener = self.ref.on('child_changed', self._process.bind(self, self._childChanged));
+    self.removeListener = self.ref.on('child_removed', self._process.bind(self, self._childRemoved));
+    
+    logger.info("Flashlight initialized type " + self.type + "; " + fbutil.getListenerCount() + " listeners.");
+    
+  })
+  .catch(function(err) {
+    logger.error("Could not preload existing data in firebase: " + err);
+  });
 };
 
+PathMonitor.prototype._stop = function() {
+  this.ref.off('child_added', this.addListener);
+  this.ref.off('child_changed', this.changeListener);
+  this.ref.off('child_removed', this.removeListener);
+};
 
-
-function nameFor(path, key) {
-  return path.index + '/' + path.type + '/' + key;
-}
-
-function parseKeys(data, fields, omit) {
-  if (!data || typeof(data)!=='object') {
-    return data;
-  }
-
-  var simpleSanitize = function(value) {
-    if (typeof(value)==='object') {
-      // Test to see if we have an "array" containing key: true refs
-      var isPseudoArray = Object.keys(value).map(function(k) {
-        return value[k];
-      }).reduce(function(a,b) { return (a===true) && (b===true); }, true);
-
-      return ( isPseudoArray ? _.keys(value) : value );
-    }
-    else if (typeof(value)==='string') {
-      return S(value).stripTags().s;
-    }
-    else {
-      return value;
-    }
-  };
-
-  // restrict to specified fields list if supplied
-  var outFields = ( Array.isArray(fields) && fields.length ? fields : Object.keys(data) );
-
-  return Q.try( function() {
-    var out = {};
-    outFields.forEach(function(f) {
-      if( data.hasOwnProperty(f) ) {
-        out[f] = simpleSanitize(data[f]);
-      }
-    });
-    // remove omitted fields
-    if( Array.isArray(omit) && omit.length) {
-      omit.forEach(function(f) {
-        if( out.hasOwnProperty(f) ) {
-          delete out[f];
-        }
-      });
-    }
-    return out;
+PathMonitor.prototype._childAdded = function(key, data) {
+  var name = this.nameFor(key);
+  this.esc.index({
+    index: this.index, 
+    type: this.type, 
+    id: key,
+    body: data
+  }).then( function(rsp) {
+    logger.debug('indexed', name);
+  }).catch( function(err) {
+    logger.error('failed to index %s: %s', name, err);
   });
-}
+};
+  
+PathMonitor.prototype._childChanged = function(key, data) {
+  var name = this.nameFor(key);
+  this.esc.index({
+    index: this.index, 
+    type: this.type, 
+    id: key,
+    body: data
+  }).then( function(rsp) {
+    logger.debug('updated', name);
+  }).catch( function(err) {
+    logger.error('failed to update %s: %s', name, err);
+  });
+};
+  
+PathMonitor.prototype._childRemoved = function(key, data) {
+  var name = this.nameFor(key);
+  this.esc.delete({
+    index: this.index, 
+    type: this.type, 
+    id: key, 
+  }).then( function(rsp) {
+    logger.debug('deleted', name);
+  }).catch( function(err) {
+    logger.error('failed to delete %s: %s', name, err);
+  });
+};
 
 exports.process = function(esc, paths, dynamicPathUrl) {
   paths && paths.forEach(function(pathProps) {
